chore: Move auxiliar method tho hub class
